backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymupdf==1.23.8
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
python-multipart==0.0.6
groq==0.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo[srv]==4.6.0
python-docx==1.1.0
email-validator==2.1.0
pydantic[email]==2.5.0
smtplib-ssl==1.0.0
